Update project with the best practices

+++ src/main/java/com/mandacarubroker/service/StockService.java
@@ -3,18 +3,20 @@ package com.mandacarubroker.service;
 import com.mandacarubroker.domain.stock.RequestStockDTO;
 import com.mandacarubroker.domain.stock.Stock;
 import com.mandacarubroker.domain.stock.StockRepository;
-import jakarta.validation.*;
-import org.springframework.stereotype.Service;
-
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.Validation;
+import jakarta.validation.ValidatorFactory;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
+import org.springframework.stereotype.Service;
 
 @Service
 public class StockService {
 
-
     private final StockRepository stockRepository;
+    private static final String NOT_FOUND_MSG = "Stock Not Found";
 
     public StockService(StockRepository stockRepository) {
         this.stockRepository = stockRepository;
@@ -24,54 +26,47 @@ public class StockService {
         return stockRepository.findAll();
     }
 
-    public Optional<Stock> getStockById(String id) {
-        return stockRepository.findById(id);
+    public Stock getStockById(String id) {
+        return stockRepository.findById(id).orElseThrow(()->new EntityNotFoundException(NOT_FOUND_MSG));
     }
 
     public Stock createStock(RequestStockDTO data) {
-        Stock novaAcao = new Stock(data);
         validateRequestStockDTO(data);
-        return stockRepository.save(novaAcao);
+        Stock newStock = new Stock(data);
+        return stockRepository.save(newStock);
     }
 
-    public Optional<Stock> updateStock(String id, Stock updatedStock) {
+    public Stock updateStock(String id, RequestStockDTO updatedStock) {
+        validateRequestStockDTO(updatedStock);
         return stockRepository.findById(id)
                 .map(stock -> {
-                    stock.setSymbol(updatedStock.getSymbol());
-                    stock.setCompanyName(updatedStock.getCompanyName());
-                    double newPrice = stock.changePrice(updatedStock.getPrice(), true);
+                    stock.setSymbol(updatedStock.symbol());
+                    stock.setCompanyName(updatedStock.companyName());
+                    double newPrice = stock.changePrice(updatedStock.price(),true);
                     stock.setPrice(newPrice);
-
                     return stockRepository.save(stock);
-                });
+                }).orElseThrow(()->new EntityNotFoundException(NOT_FOUND_MSG));
     }
 
     public void deleteStock(String id) {
+        if (!stockRepository.existsById(id)){
+            throw new EntityNotFoundException(NOT_FOUND_MSG);
+        }
         stockRepository.deleteById(id);
     }
 
     public static void validateRequestStockDTO(RequestStockDTO data) {
         ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        Validator validator = factory.getValidator();
-        Set<ConstraintViolation<RequestStockDTO>> violations = validator.validate(data);
-
+        Set<ConstraintViolation<RequestStockDTO>> violations = factory.getValidator().validate(data);
         if (!violations.isEmpty()) {
             StringBuilder errorMessage = new StringBuilder("Validation failed. Details: ");
-
             for (ConstraintViolation<RequestStockDTO> violation : violations) {
                 errorMessage.append(String.format("[%s: %s], ", violation.getPropertyPath(), violation.getMessage()));
             }
-
             errorMessage.delete(errorMessage.length() - 2, errorMessage.length());
-
+            factory.close();
             throw new ConstraintViolationException(errorMessage.toString(), violations);
         }
-    }
-
-    public void validateAndCreateStock(RequestStockDTO data) {
-        validateRequestStockDTO(data);
-
-        Stock novaAcao = new Stock(data);
-        stockRepository.save(novaAcao);
+        factory.close();
     }
 }
